requirements.txt
fastapi
uvicorn
telethon
python-dotenv
motor
langid
apscheduler
chromadb
